fix load audio for tfio.audio

+++ vistec_ser/data/features/preprocessing.py
@@ -17,7 +17,7 @@ def load_waveform(
     #     waveform, _ = tf.audio.decode_wav(audio_binary)
     # elif '.flac' in audio:
     #     waveform = tf.divide(tfio.audio.decode_flac(audio_binary, dtype=tf.int16), normalize_factor)
-    audio_tensor = tfio.audio.AudioIOTensor(filename=audio, dtype=tf.float32)
+    audio_tensor = tfio.audio.AudioIOTensor(filename=audio, dtype=tf.int64)
     audio_rate = audio_tensor.rate
     waveform = tf.divide(audio_tensor.to_tensor(), normalize_factor)
     waveform = tf.squeeze(waveform, axis=-1)
